Code Improvements after Review
